Rename xen_types.rs to domctl_types.rs
