+++ crates/xen-ioctls/src/domctl_types.rs
